\name{VaRES-package}
\alias{VaRES-package}
\alias{VaRES}
\docType{package}
\title{
Computes value at risk and expected shortfall for over 100 parametric distributions
}
\description{Computes Value at risk and expected shortfall, two most popular measures of financial risk, for over one hundred parametric distributions,  including all commonly known distributions.  Also computed are the corresponding probability density function and cumulative distribution function.}

\references{Stephen Chan, Saralees Nadarajah & Emmanuel Afuecheta (2016). An R Package for Value at Risk and Expected Shortfall, Communications in Statistics - Simulation and Computation, 45:9, 3416-3434, \doi{10.1080/03610918.2014.944658}}
\keyword{package}
