\name{VaRES-package}
\alias{VaRES-package}
\alias{VaRES}
\docType{package}
\title{
Computes value at risk and expected shortfall for over 100 parametric distributions
}
\description{Computes Value at risk and expected shortfall, two most popular measures of financial risk, for over one hundred parametric distributions,  including all commonly known distributions.  Also computed are the corresponding probability density function and cumulative distribution function.}
\details{
\tabular{ll}{
Package: \tab VaRES\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-8-25\cr
License: \tab GPL(>=2)
}}
\author{Saralees Nadarajah, Stephen Chan and Emmanuel Afuecheta

Maintainer: Saralees Nadarajah <Saralees.Nadarajah@manchester.ac.uk>
}
\references{S. Nadarajah, S. Chan and E. Afuecheta, An R Package for value at risk and expected shortfall, submitted}
\keyword{package}
